+++ CHANGELOG.md
@@ -26,6 +26,7 @@
 - [FIX] Now search for devices by Mac address and by flipper name
 - [FIX] Subghz provisioning while geoip is null
 - [FIX] Separate screen for edit key
+- [FIX] Synchronization with shadow files
 
 # 1.2.0